// backend/src/common/enums/order-status.enum.ts
export enum OrderStatus {
  AWAITING_PREPAYMENT = 'AWAITING_PREPAYMENT',
  PENDING_MATCH = 'PENDING_MATCH',
  OFFERED = 'OFFERED',
  ACCEPTED = 'ACCEPTED',
  ARRIVED = 'ARRIVED',
  IN_PROGRESS = 'IN_PROGRESS',
  AWAITING_PRICE_APPROVAL = 'AWAITING_PRICE_APPROVAL',
  PRICE_DISPUTED = 'PRICE_DISPUTED',
  AWAITING_PAYMENT = 'AWAITING_PAYMENT',
  PAID = 'PAID',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  REFUND_PENDING = 'REFUND_PENDING',
  REFUNDED = 'REFUNDED',
}